Package structure: moved UI utilities.
Part of #963.
