+++ src/main/java/de/dennisguse/opentracks/ui/markers/MarkerResourceCursorAdapter.java
@@ -23,11 +23,11 @@ import java.time.Instant;
 import java.time.OffsetDateTime;
 
 import de.dennisguse.opentracks.R;
-import de.dennisguse.opentracks.adapters.ExecutorListViewService;
-import de.dennisguse.opentracks.adapters.ScrollVisibleViews;
 import de.dennisguse.opentracks.content.data.MarkerColumns;
 import de.dennisguse.opentracks.content.data.Track;
 import de.dennisguse.opentracks.content.provider.ContentProviderUtils;
+import de.dennisguse.opentracks.ui.util.ExecutorListViewService;
+import de.dennisguse.opentracks.ui.util.ScrollVisibleViews;
 import de.dennisguse.opentracks.util.ListItemUtils;
 
 public class MarkerResourceCursorAdapter extends ResourceCursorAdapter implements ScrollVisibleViews.VisibleViewsListener {
